changed shopfront button
