+++ src/components/Shopfront/Shopfront.jsx
@@ -1,19 +1,10 @@
-import React, { useState } from 'react'
+import React from 'react'
 import '../styles/cartItem-styles.css'
-import { useCartContext } from '../../app-context/CartContext'
 import ShopfrontItem from './ShopfrontItem'
 
 
 
 const ShopFront = ({ items }) => {
-
-  const { addToCart } = useCartContext()
-  const [buttonText, setButtonText] = useState("Add to cart")
-
-  const handleClick = (item) => {
-    setButtonText(buttonText === 'Add to cart' ? 'Added to cart!' : 'Add to cart')
-    addToCart(item)
-  }
 
   return (
     <>
@@ -24,12 +15,11 @@ const ShopFront = ({ items }) => {
         return (
           <ShopfrontItem
           key={item.id}
+          item={item}
           name={item.name}
           price={item.price}
           description={item.description}
           image={item.image}
-          buttonText={buttonText}
-          onAdd={()=> handleClick(item)}
           /> 
         )
         })
